+++ metadata.rb
@@ -6,11 +6,10 @@ description       'Installs/Configures karaf'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 source_url        'https://github.com/jcapriotti/apache-karaf-cookbook'
 issues_url        'https://github.com/JCapriotti/apache-karaf-cookbook/issues'
-version           '0.2.1'
+version           '1.0.0'
 
 depends 'ark'
 depends 'java'
 
 supports 'centos'
-supports 'fedora', '<= 21'
 supports 'ubuntu'