+++ resources/karaf_feature.rb
@@ -1,0 +1,29 @@
+resource_name :karaf_feature
+
+property :install_path, kind_of: String, default: '/usr/local'
+property :client_user, kind_of: String, default: 'karaf'
+property :feature_name, kind_of: String, required: true, name_property: true
+property :version, kind_of: String, default: ''
+
+default_action :install
+
+client_command = "bin/client"
+def karaf_path
+  "#{install_path}/karaf"
+end 
+
+action :install do
+
+  if version.empty?
+    feature_string = feature_name
+  else
+    feature_string = "#{feature_name}/#{version}"
+  end
+
+  execute 'install feature' do
+    cwd  karaf_path
+    command <<-EOF
+      #{client_command} -u karaf feature:install #{feature_string}
+    EOF
+  end
+end