+++ test/cookbooks/karaf_test/recipes/default.rb
@@ -1,0 +1,27 @@
+
+
+karaf 'install karaf' do
+  install_java  true
+  version       '3.0.3'
+  action        :install
+end
+
+karaf_feature_repository 'hawtio' do
+  version '1.4.51'
+  :install
+end
+
+karaf_feature 'hawtio' do
+  :install
+end
+
+karaf_user 'newuser' do
+  groups    ['group1', 'group2']
+  password  'ultrafubar'
+  :create
+end
+
+karaf_group 'newgroup' do
+  roles   ['rrrrrrrrrrrrr1', 'rrrrrrrrrrrr2']
+  :create
+end